requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
matplotlib
shap
requests
supabase-py
Faker
seaborn
xgboost
